+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 # main
 
+* Added cop `RSpec::SidekiqInline` ([#24](https://github.com/petalmd/rubocop-petal/pull/24))
 * Remove cop `Rails/ValidateUniquenessCase` ([#21](https://github.com/petalmd/rubocop-petal/pull/21))
 
 # v0.5.0